split scene further
