--- src/scene/.#geo.rs
@@ -1,1 +1,0 @@
-ajay@Stephens-MacBook-Pro-2.local.36812
