+++ src/scene/mod.rs
@@ -1,265 +1,12 @@
-use std::ops::{Add, Sub};
+extern crate sdl2;
+
+mod geo;
+mod objects;
+
+pub use geo::{Point, Ray, Vector};
+use objects::{Material, Object, Plane, Sphere, BSDF};
 
 use crate::common::{Spectrum, EPS, GENERIC_ERROR};
-
-#[derive(Clone, Copy, Debug)]
-pub struct Ray {
-    pub origin: Point,
-    pub direction: Vector,
-    pub bounces_left: u32,
-}
-
-impl Ray {
-    // normalizes the direction vector
-    pub fn new(origin: Point, direction: Vector, bounces_left: u32) -> Ray {
-        Ray {
-            origin,
-            direction: direction.normalized(),
-            bounces_left,
-        }
-    }
-
-    pub fn get_intersection_point(&self, scalar: f64) -> Point {
-        let scaled_vector = self.direction.scale(scalar);
-        self.origin + scaled_vector
-    }
-}
-
-#[derive(Clone, Copy, Debug)]
-pub struct Point {
-    pub x: f64,
-    pub y: f64,
-    pub z: f64,
-}
-
-impl Point {
-    pub fn new(x: f64, y: f64, z: f64) -> Point {
-        Point { x, y, z }
-    }
-}
-
-#[derive(Clone, Copy, Debug)]
-pub struct Vector {
-    pub x: f64,
-    pub y: f64,
-    pub z: f64,
-}
-
-fn norm(x: f64, y: f64, z: f64) -> f64 {
-    f64::sqrt(f64::powi(x, 2) + f64::powi(y, 2) + f64::powi(z, 2))
-}
-
-impl Vector {
-    pub fn new(x: f64, y: f64, z: f64) -> Vector {
-        Vector { x, y, z }
-    }
-
-    pub fn new_normalized(x: f64, y: f64, z: f64) -> Vector {
-        let mut vector = Vector::new(x, y, z);
-        vector.normalize();
-        vector
-    }
-
-    pub fn norm(&self) -> f64 {
-        norm(self.x, self.y, self.z)
-    }
-
-    pub fn normalize(&mut self) {
-        let inverse_norm = 1.0 / self.norm();
-        self.x *= inverse_norm;
-        self.y *= inverse_norm;
-        self.z *= inverse_norm;
-    }
-
-    pub fn normalized(&self) -> Vector {
-        let mut return_vector = *self;
-        return_vector.normalize();
-        return_vector
-    }
-
-    pub fn dot(&self, other_vector: Vector) -> f64 {
-        self.x * other_vector.x + self.y * other_vector.y + self.z * other_vector.z
-    }
-
-    pub fn scale(&self, scalar: f64) -> Vector {
-        Vector::new(scalar * self.x, scalar * self.y, scalar * self.z)
-    }
-}
-
-impl Sub for Point {
-    type Output = Vector;
-
-    fn sub(self, other: Point) -> Self::Output {
-        Vector::new(self.x - other.x, self.y - other.y, self.z - other.z)
-    }
-}
-
-impl Add<Vector> for Point {
-    type Output = Point;
-
-    fn add(self, other: Vector) -> Self::Output {
-        Point::new(self.x + other.x, self.y + other.y, self.z + other.z)
-    }
-}
-
-impl Add for Vector {
-    type Output = Vector;
-
-    fn add(self, other: Vector) -> Self::Output {
-        Vector::new(self.x + other.x, self.y + other.y, self.z + other.z)
-    }
-}
-
-impl Sub for Vector {
-    type Output = Vector;
-
-    fn sub(self, other: Vector) -> Self::Output {
-        Vector::new(self.x - other.x, self.y - other.y, self.z - other.z)
-    }
-}
-
-#[derive(Clone, Copy, Debug)]
-pub enum BSDF {
-    Diffuse,
-}
-
-#[derive(Clone, Copy, Debug)]
-pub struct Material {
-    pub bsdf: BSDF,
-    pub reflectance: Spectrum,
-    pub emittance: Spectrum,
-}
-
-pub trait Object {
-    fn intersect(&self, ray: &Ray) -> Option<f64>;
-    fn surface_normal(&self, point: Point) -> Vector;
-    fn material(&self) -> &Material;
-}
-
-pub struct Sphere {
-    pub center: Point,
-    pub radius: f64,
-    material: Material,
-}
-
-pub struct Plane {
-    pub point: Point,
-    pub normal: Vector,
-    material: Material,
-}
-
-impl Material {
-    pub fn new(bsdf: BSDF, reflectance: Spectrum, emittance: Spectrum) -> Material {
-        Material {
-            bsdf,
-            reflectance,
-            emittance,
-        }
-    }
-
-    pub fn sample_bsdf(&self, wi: Vector, wo: Vector) -> Spectrum {
-        match self.bsdf {
-            BSDF::Diffuse => self.reflectance * (1.0 / std::f64::consts::PI),
-        }
-    }
-
-    pub fn bounce(&self, wi: Vector) -> Vector {
-        match self.bsdf {
-            BSDF::Diffuse => {
-                // return random vector
-            }
-        }
-        unimplemented!();
-    }
-}
-
-impl Sphere {
-    pub fn new(center: Point, radius: f64, material: Material) -> Sphere {
-        Sphere {
-            center,
-            radius,
-            material,
-        }
-    }
-}
-
-impl Object for Sphere {
-    // sphere intersection from bheisler
-    // returns intersection distance
-    fn intersect(&self, ray: &Ray) -> Option<f64> {
-        let l: Vector = self.center - ray.origin;
-        let adj = l.dot(ray.direction);
-        let d2 = l.dot(l) - (adj * adj);
-        let radius2 = self.radius * self.radius;
-        if d2 > radius2 {
-            return None;
-        }
-        let thc = (radius2 - d2).sqrt();
-        let t0 = adj - thc;
-        let t1 = adj + thc;
-
-        if t0 < 0.0 && t1 < 0.0 {
-            return None;
-        }
-
-        let distance = if t0 < t1 { t0 } else { t1 };
-        Some(distance)
-    }
-
-    fn surface_normal(&self, point: Point) -> Vector {
-        (point - self.center).normalized()
-    }
-
-    fn material(&self) -> &Material {
-        &self.material
-    }
-}
-
-impl Plane {
-    pub fn new(point: Point, normal: Vector, material: Material) -> Plane {
-        Plane {
-            point,
-            normal,
-            material,
-        }
-    }
-}
-
-impl Object for Plane {
-    fn intersect(&self, ray: &Ray) -> Option<f64> {
-        let d = (self.point - ray.origin).dot(self.normal) / ray.direction.dot(self.normal);
-        if d > 0.0 {
-            Some(d)
-        } else {
-            None
-        }
-    }
-
-    fn surface_normal(&self, _point: Point) -> Vector {
-        self.normal
-    }
-
-    fn material(&self) -> &Material {
-        &self.material
-    }
-}
-
-// #[cfg(test)]
-// mod tests {
-//     use super::*;
-//     #[test]
-//     fn sphere_intersection() {
-//         let origin = Point::new(0.0, 0.0, 0.0);
-//         let direction = Vector::new_normalized(0.0, 0.0, -1.0);
-//         let ray = Ray::new(&origin, &direction);
-//         let sphere = Sphere::new(Point::new(0.0, 0.0, -4.0), 2.0);
-//         assert!(sphere.intersect(&ray) != None);
-
-//         let vector_that_misses = Vector::new_normalized(3.0, 0.0, -4.0);
-//         let ray_that_misses = Ray::new(&origin, &vector_that_misses);
-//         assert!(sphere.intersect(&ray_that_misses) == None);
-//     }
-// }
 
 pub struct Scene {
     planes: Vec<Plane>,
